pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pmc-retrieval"
version = "0.1.0"
description = "PMC retrieval service with OpenSearch and PostgreSQL"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "PyYAML>=6.0.1",
  "opensearch-py>=2.6.0",
  "lark>=1.2.2",
  "psycopg[binary]>=3.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*", "parser*", "search*", "ingest*", "storage*"]
